util: Optimize and add some utility functions
Change input parameter of wait function to a 64-bit cycle count instead of 32-bits
Add bit splicing operations to util.h for easily managing bit fields within a register
Update read32 and write32 to get rid of unnecessary stack store instructions

+++ include/time.h
@@ -5,6 +5,6 @@
 
 #include "types.h"
 
-void wait(u32 cycles);
+void wait(u64 cycles);
 
 #endif
